This is Grading system application of Resturant.By using this app people can making sure their safety on food
Food safety is too much important in our country.Most of the people in our city often go to resturant for different perposes like birthday,marriage cremony,celebrating many of other reason and in our country rasturant business is getting popular day by day.But now a days resturant authority don't care about the maintenance  the rules of safety food or hygenic food.
So we try to grading that resturant for our own safety,so that we may getting hygenice food.